fix: 修复square double类型

+++ lib/service/backend_service/select_house/house_page_entity.dart
@@ -87,7 +87,7 @@ class Content {
   int? price;
   String? source;
   String? residential;
-  int? squares;
+  double? squares;
   String? layout;
   int? shi;
   int? ting;
